+++ packages/core/engine/index.ts
@@ -1,6 +1,6 @@
-export * from './action.js';
-export * as GameConfig from './GameConfig.js';
+export * from './GameConfig.js';
 export * from './PlayerData.js';
+export * from './action.js';
 export * as Socket from './socket.js';
 export * from './utils.js';