+++ packages/core/index.ts
@@ -2,9 +2,9 @@ export * from './common/utils.js';
 export * from './constant/index.js';
 
 export * from './battle/index.js';
+export * from './emitter/index.js';
 export * from './engine/index.js';
 export * from './entity/index.js';
-export * from './event-bus/index.js';
 export * from './level/index.js';
 
 export * from './functions/index.js';