+++ packages/launcher/src/App.tsx
@@ -5,16 +5,16 @@ import { ThemeProvider } from '@mui/system';
 
 import { SEAContext } from '@sea-launcher/context/SAContext';
 
-import { SeaModuleLogger } from '@sea-launcher/utils/logger';
+import { SEAModuleLogger } from '@sea-launcher/utils/logger';
 
 import { saTheme } from '@sea-launcher/style';
 
 import { CommandBar } from './CommandBar';
 import { MainPanel } from './views/MainPanel';
 
-import { Hook } from 'sea-core';
+import { Hook, SEAHookDispatcher as HookDispatcher } from 'sea-core';
 import { Manager as BattleManager, autoStrategy } from 'sea-core/battle';
-import { EventBus, SocketListener, type SocketReceiveHandler, type SocketSendHandler } from 'sea-core/event-bus';
+import { EventBus, SocketEventEmitter } from '../../core/dist/emitter';
 
 import { QuickAccess } from './QuickAccess';
 import { SEAModManager } from './service/ModManager';
@@ -24,9 +24,9 @@ const eventBus = new EventBus();
 // console.log(import.meta.env.DEV);
 
 const Logger = {
-    BattleManager: { info: SeaModuleLogger('BattleManager', 'info') },
-    AwardManager: { info: SeaModuleLogger('AwardManager', 'info') },
-    ModuleManger: { info: SeaModuleLogger('ModuleManger', 'info') },
+    BattleManager: { info: SEAModuleLogger('BattleManager', 'info') },
+    AwardManager: { info: SEAModuleLogger('AwardManager', 'info') },
+    ModuleManger: { info: SEAModuleLogger('ModuleManger', 'info') },
 };
 
 export default function App() {
@@ -57,25 +57,29 @@ export default function App() {
 
         document.body.addEventListener('keydown', handleShortCut);
 
-        eventBus.hook(Hook.Battle.battleStart, () => {
+        const SEAHookDispatcher = eventBus.proxy(HookDispatcher);
+        const SEASocketListener = eventBus.proxy(SocketEventEmitter);
+
+        SEAHookDispatcher.on(Hook.Battle.battleStart, () => {
             toggleFighting(true);
         });
-        eventBus.hook(Hook.Battle.battleStart, handleBattleRoundEnd);
-        eventBus.hook(Hook.Battle.roundEnd, handleBattleRoundEnd);
-        eventBus.hook(Hook.Battle.battleEnd, () => {
+
+        SEAHookDispatcher.on(Hook.Battle.battleStart, handleBattleRoundEnd);
+        SEAHookDispatcher.on(Hook.Battle.roundEnd, handleBattleRoundEnd);
+        SEAHookDispatcher.on(Hook.Battle.battleEnd, () => {
             toggleFighting(false);
         });
 
-        eventBus.hook(Hook.Battle.battleStart, () => {
+        SEAHookDispatcher.on(Hook.Battle.battleStart, () => {
             Logger.BattleManager.info(`检测到对战开始`);
         });
 
-        eventBus.hook(Hook.Battle.battleEnd, () => {
+        SEAHookDispatcher.on(Hook.Battle.battleEnd, () => {
             const win = Boolean(FightManager.isWin);
             Logger.BattleManager.info(`检测到对战结束 对战胜利: ${win}`);
         });
 
-        eventBus.hook(Hook.Award.receive, (data) => {
+        SEAHookDispatcher.on(Hook.Award.receive, (data) => {
             Logger.AwardManager.info(`获得物品:`);
             const logStr = Array.isArray(data.items)
                 ? data.items.map((v) => `${ItemXMLInfo.getName(v.id)} ${v.count}`)
@@ -83,15 +87,15 @@ export default function App() {
             logStr && Logger.AwardManager.info(logStr.join('\r\n'));
         });
 
-        eventBus.hook(Hook.Module.loadScript, (name) => {
+        SEAHookDispatcher.on(Hook.Module.loadScript, (name) => {
             Logger.ModuleManger.info(`检测到新模块加载: ${name}`);
         });
 
-        eventBus.hook(Hook.Module.openMainPanel, ({ module, panel }) => {
+        SEAHookDispatcher.on(Hook.Module.openMainPanel, ({ module, panel }) => {
             Logger.ModuleManger.info(`${module}创建主面板: ${panel}`);
         });
 
-        const noteUseSkillHandler: SocketReceiveHandler<typeof CommandID.NOTE_USE_SKILL> = (data) => {
+        SEASocketListener.on(CommandID.NOTE_USE_SKILL, 'receive', (data) => {
             const [fi, si] = data;
             Logger.BattleManager.info(`对局信息更新:
                 先手方:${fi.userId}
@@ -105,17 +109,14 @@ export default function App() {
                 造成伤害: ${si.damage}
                 是否暴击:${si.isCrit}
                 使用技能: ${SkillXMLInfo.getName(si.skillId)}`);
-        };
+        });
 
-        const useSkillHandler: SocketSendHandler = (data) => {
+        SEASocketListener.on(CommandID.USE_SKILL, 'send', (data) => {
             const [skillId] = data as [number];
             Logger.BattleManager.info(
                 `${FighterModelFactory.playerMode?.info.petName} 使用技能: ${SkillXMLInfo.getName(skillId)}`
             );
-        };
-
-        SocketListener.on(CommandID.NOTE_USE_SKILL, 'receive', noteUseSkillHandler);
-        SocketListener.on(CommandID.USE_SKILL, 'send', useSkillHandler);
+        });
 
         let active = true;
         SEAModManager.fetchMods().then((mods) => {
@@ -127,8 +128,6 @@ export default function App() {
 
         const clean = () => {
             active = false;
-            SocketListener.off(CommandID.NOTE_USE_SKILL, 'receive', noteUseSkillHandler);
-            SocketListener.off(CommandID.USE_SKILL, 'send', useSkillHandler);
             SEAModManager.teardown();
             eventBus.unmount();
             document.body.removeEventListener('keydown', handleShortCut);