+++ packages/launcher/src/views/Realm/LevelElfKingsTrial.tsx
@@ -2,7 +2,7 @@ import { LevelState, Socket } from 'sea-core';
 
 import type { ILevelBattleStrategy, ILevelRunner, LevelData as SEALevelData, LevelInfo as SEALevelInfo } from 'sea-core';
 
-import { SeaModuleLogger } from '@sea-launcher/utils/logger';
+import { SEAModuleLogger } from '@sea-launcher/utils/logger';
 import dataProvider from './data';
 
 const customData = dataProvider['LevelElfKingsTrial'];
@@ -51,7 +51,7 @@ export class LevelElfKingsTrial implements ILevelRunner<LevelData, SEALevelInfo>
 
     option: LevelOption;
 
-    logger = SeaModuleLogger('精灵王的试炼', 'info');
+    logger = SEAModuleLogger('精灵王的试炼', 'info');
 
     constructor(option: LevelOption) {
         this.option = option;