+++ packages/launcher/src/views/Realm/LevelExpTraining.tsx
@@ -2,7 +2,7 @@ import { LevelState, Socket } from 'sea-core';
 
 import type { ILevelBattleStrategy, ILevelRunner, LevelData as SEALevelData, LevelInfo as SEALevelInfo } from 'sea-core';
 
-import { SeaModuleLogger } from '@sea-launcher/utils/logger';
+import { SEAModuleLogger } from '@sea-launcher/utils/logger';
 import dataProvider from './data';
 
 const customData = dataProvider['LevelExpTraining'];
@@ -33,7 +33,7 @@ export class LevelExpTraining implements ILevelRunner<LevelData, SEALevelInfo> {
 
     option: LevelOption;
 
-    logger = SeaModuleLogger('经验训练场', 'info');
+    logger = SEAModuleLogger('经验训练场', 'info');
 
     constructor(option: LevelOption) {
         this.option = option;